+++ app/src/adapter/nextjs/nextjs.adapter.ts
@@ -1,7 +1,7 @@
 import type { IncomingMessage, ServerResponse } from "node:http";
+import { nodeRequestToRequest } from "adapter/utils";
 import type { App } from "bknd";
 import { type FrameworkBkndConfig, createFrameworkApp } from "bknd/adapter";
-import { nodeRequestToRequest } from "bknd/adapter/node";
 import { Api } from "bknd/client";
 
 export type NextjsBkndConfig = FrameworkBkndConfig & {
